feat: new temporal API

+++ src/Reminder.ts
@@ -14,7 +14,7 @@ import { rest } from "./util/globals";
 export type Params = {
 	message: RESTPostAPIChannelMessageJSONBody;
 	remind: string;
-	duration: number;
+	timestamp: number;
 	userId: string;
 };
 
@@ -23,15 +23,14 @@ export class Reminder extends WorkflowEntrypoint<Env, Params> {
 		event: Readonly<WorkflowEvent<Params>>,
 		step: WorkflowStep,
 	) {
-		const sleep = step.sleep("Sleep", event.payload.duration);
-
 		await step.do<void>("Store reminder", this.storeReminder.bind(this, event));
 		const channelId = await step.do(
 			"Create dm channel",
 			this.createDM.bind(this, event.payload.userId),
 		);
 
-		await sleep;
+		if (event.payload.timestamp > Date.now())
+			await step.sleepUntil("Sleep", new Date(event.payload.timestamp));
 		await Promise.all([
 			step.do<void>(
 				"Send reminder",
@@ -43,18 +42,15 @@ export class Reminder extends WorkflowEntrypoint<Env, Params> {
 
 	private async storeReminder({
 		instanceId,
-		payload: { duration, remind, userId },
+		payload: { timestamp, remind, userId },
 	}: WorkflowEvent<Params>) {
 		await this.env.DB.prepare(
-			`INSERT INTO Reminders (id, date, userId, remind)
-				VALUES (?1, datetime('now', '+' || ?2 || ' seconds'), ?3, ?4)`,
+			`
+				INSERT INTO Reminders (id, timestamp, userId, remind)
+				VALUES (?1, ?2, ?3, ?4)
+			`,
 		)
-			.bind(
-				instanceId.slice(userId.length + 1),
-				Math.ceil(duration / 1_000),
-				userId,
-				remind,
-			)
+			.bind(instanceId.slice(userId.length + 1), timestamp, userId, remind)
 			.run();
 	}
 
@@ -80,8 +76,7 @@ export class Reminder extends WorkflowEntrypoint<Env, Params> {
 		payload: { userId },
 	}: WorkflowEvent<Params>) {
 		await this.env.DB.prepare(
-			`DELETE FROM Reminders
-			WHERE id = ?1 AND userId = ?2`,
+			`DELETE FROM Reminders WHERE id = ?1 AND userId = ?2`,
 		)
 			.bind(instanceId.slice(userId.length + 1), userId)
 			.run();
