+++ src/commands/remind.ts
@@ -1,4 +1,4 @@
-import { env } from "cloudflare:workers";
+import { env, waitUntil } from "cloudflare:workers";
 import {
 	ApplicationCommandOptionType,
 	ApplicationCommandType,
@@ -9,10 +9,11 @@ import {
 	type APIComponentInContainer,
 	type RESTPostAPIApplicationCommandsJSONBody,
 } from "discord-api-types/v10";
+import { Temporal } from "temporal-polyfill";
 import Command from "../Command";
 import normalizeError from "../util/normalizeError";
 import { maxLength } from "../util/strings";
-import { parseTime, TimeUnit } from "../util/time";
+import { idToTimestamp, parseDuration } from "../util/time";
 
 export class Remind extends Command {
 	static override chatInputData = {
@@ -37,7 +38,11 @@ export class Remind extends Command {
 						description: "Tra quanto tempo te lo dovrò ricordare (es. 1d3h)",
 						type: ApplicationCommandOptionType.String,
 						required: true,
-						max_length: 1024,
+					},
+					{
+						name: "tz",
+						description: "Il fuso orario (default Europe/Rome)",
+						type: ApplicationCommandOptionType.String,
 					},
 				],
 			},
@@ -70,35 +75,40 @@ export class Remind extends Command {
 		{
 			options,
 			user: { id: userId },
+			interaction: { id: interactionId },
 		}: ChatInputArgs<typeof Remind.chatInputData, "me">,
 	) => {
-		const duration = parseTime(options.in);
-		if (Number.isNaN(duration))
+		let start;
+		try {
+			start = Temporal.Instant.fromEpochMilliseconds(
+				idToTimestamp(interactionId),
+			).toZonedDateTimeISO(options.tz ?? "Europe/Rome");
+		} catch (err) {
+			return reply({
+				content: "Il fuso orario inserito non è valido!",
+				flags: MessageFlags.Ephemeral,
+			});
+		}
+		const target = parseDuration(options.in, start);
+		if (!target)
 			return reply({
 				content: "La durata inserita non è valida!",
 				flags: MessageFlags.Ephemeral,
 			});
-		if (duration > TimeUnit.Year)
+		if (Temporal.ZonedDateTime.compare(target, start.add({ seconds: 5 })) <= 0)
 			return reply({
-				content: "Non puoi impostare una durata maggiore di 1 anno!",
+				content: `La durata inserita è troppo breve (<t:${Math.round(target.epochMilliseconds / 1000)}:R>)!`,
 				flags: MessageFlags.Ephemeral,
 			});
-		if (duration < TimeUnit.Second)
+		if (Temporal.ZonedDateTime.compare(target, start.add({ years: 4 })) > 0)
 			return reply({
-				content: `Non puoi impostare una durata minore di 1 secondo (<t:${Math.round((Date.now() + duration) / 1_000)}:R>)!`,
-				flags: MessageFlags.Ephemeral,
-			});
-		if (options.to.length > 1_000)
-			return reply({
-				content: "La lunghezza massima di un promemoria è di 1000 caratteri",
+				content: `La durata inserita è troppo lunga (<t:${Math.round(target.epochMilliseconds / 1000)}:R>)!`,
 				flags: MessageFlags.Ephemeral,
 			});
 		defer({ flags: MessageFlags.Ephemeral });
 		if (
 			((await env.DB.prepare(
-				`SELECT COUNT(*) as count
-						FROM Reminders
-						WHERE userId = ?1`,
+				`SELECT COUNT(*) as count FROM Reminders WHERE userId = ?1`,
 			)
 				.bind(userId)
 				.first<number>("count")) ?? 0) >= 64
@@ -109,7 +119,7 @@ export class Remind extends Command {
 			id: `${userId}-${id}`,
 			params: {
 				message: { content: `🔔 Promemoria: ${options.to}` },
-				duration,
+				timestamp: target.epochMilliseconds,
 				userId,
 				remind: options.to,
 			},
@@ -120,7 +130,7 @@ export class Remind extends Command {
 				content: `Si è verificato un errore: \`${result.message}\``,
 			});
 		return edit({
-			content: `Promemoria \`${id}\` impostato correttamente <t:${Math.round((Date.now() + duration) / 1_000)}:R>!`,
+			content: `Promemoria \`${id}\` impostato correttamente <t:${Math.round(target.epochMilliseconds / 1000)}:R>!`,
 		});
 	};
 	static list = (
@@ -136,11 +146,18 @@ export class Remind extends Command {
 		return this.sendPage(edit, userId, Number(page) || 0);
 	};
 	static remove = async (
-		{ defer, reply, edit }: ChatInputReplies | ComponentReplies,
 		{
-			args: [, argId] = [],
+			defer,
+			reply,
+			edit,
+			deferUpdate,
+			followup,
+		}: Merge<ChatInputReplies, ComponentReplies>,
+		{
+			args: [, argId, page] = [],
 			options: { id } = { id: argId! },
 			user: { id: userId },
+			interaction,
 		}: Merge<
 			ChatInputArgs<typeof Remind.chatInputData, "remove">,
 			ComponentArgs
@@ -152,13 +169,11 @@ export class Remind extends Command {
 				content: "Promemoria non trovato!",
 				flags: MessageFlags.Ephemeral,
 			});
-		defer({ flags: MessageFlags.Ephemeral });
+		if (interaction.type === InteractionType.MessageComponent) deferUpdate!();
+		else defer({ flags: MessageFlags.Ephemeral });
 		const error = await Promise.all([
 			instance.terminate(),
-			env.DB.prepare(
-				`DELETE FROM Reminders
-					WHERE id = ?1 AND userId = ?2`,
-			)
+			env.DB.prepare(`DELETE FROM Reminders WHERE id = ?1 AND userId = ?2`)
 				.bind(id, userId)
 				.run(),
 		])
@@ -166,10 +181,22 @@ export class Remind extends Command {
 			.catch(normalizeError);
 
 		if (error)
-			return edit({
+			return (
+				interaction.type === InteractionType.MessageComponent ?
+					followup
+				:	edit)({
 				content: `Si è verificato un errore: \`${error.message}\``,
+				flags: MessageFlags.Ephemeral,
 			});
-		return edit({ content: "Promemoria rimosso correttamente!" });
+		if (interaction.type === InteractionType.MessageComponent)
+			waitUntil(this.sendPage(edit, userId, Number(page) || 0));
+		return (
+			interaction.type === InteractionType.MessageComponent ?
+				followup
+			:	edit)({
+			content: "Promemoria rimosso correttamente!",
+			flags: MessageFlags.Ephemeral,
+		});
 	};
 	static sendPage = async (
 		edit: ChatInputReplies["edit"],
@@ -178,13 +205,15 @@ export class Remind extends Command {
 	) => {
 		const { results } = await env.DB.prepare(
 			`
-				SELECT id, date, remind, COUNT(*) OVER () AS count
-				FROM Reminders WHERE userId = ?1 ORDER BY date LIMIT 8 OFFSET ?2
+				SELECT id, timestamp, remind, COUNT(*) OVER () AS count
+				FROM Reminders WHERE userId = ?1 ORDER BY timestamp LIMIT 10 OFFSET ?2
 			`,
 		)
-			.bind(userId, page * 8)
+			.bind(userId, page * 10)
 			.run<
-				Pick<Database.Reminder, "date" | "id" | "remind"> & { count: number }
+				Pick<Database.Reminder, "timestamp" | "id" | "remind"> & {
+					count: number;
+				}
 			>();
 		const count = results[0]?.count ?? 0;
 
@@ -202,14 +231,14 @@ export class Remind extends Command {
 								},
 								{ type: ComponentType.Separator },
 								...results.map((r, i) => {
-									const date = Math.round(Date.parse(`${r.date}Z`) / 1000);
+									const date = Math.round(r.timestamp / 1000);
 
 									return {
 										type: ComponentType.Section,
 										components: [
 											{
 												type: ComponentType.TextDisplay,
-												content: `${i + 1 + page * 8}. \`${r.id}\` <t:${date}:F> (<t:${date}:R>)\n>>> ${maxLength(
+												content: `${i + 1 + page * 10}. \`${r.id}\` <t:${date}:F> (<t:${date}:R>)\n>>> ${maxLength(
 													r.remind,
 													148,
 												)}`,
@@ -219,7 +248,7 @@ export class Remind extends Command {
 											type: ComponentType.Button,
 											label: "Elimina",
 											style: ButtonStyle.Danger,
-											custom_id: `remind-remove-${userId}-${r.id}`,
+											custom_id: `remind-remove-${userId}-${r.id}-${page}`,
 										},
 									} satisfies APIComponentInContainer;
 								}),
@@ -236,13 +265,13 @@ export class Remind extends Command {
 										{
 											custom_id: `remind-list-${userId}-${page}`,
 											disabled: true,
-											label: `Pagina ${page + 1} di ${Math.ceil(count / 8)}`,
+											label: `Pagina ${page + 1} di ${Math.ceil(count / 10)}`,
 											style: ButtonStyle.Secondary,
 											type: ComponentType.Button,
 										},
 										{
 											custom_id: `remind-list-${userId}-${page + 1}`,
-											disabled: count <= (page + 1) * 8,
+											disabled: count <= (page + 1) * 10,
 											emoji: { name: "➡️" },
 											style: ButtonStyle.Primary,
 											type: ComponentType.Button,
